refactor: reorganize pages and navbars
