--- src/components/ErrorPage.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-export default function ErrorPage() {
-  return (
-    <div className="h-screen flex items-center justify-center">
-      <h1 className="text-2xl font-semibold text-red-600">Something went wrong. Please try again.</h1>
-    </div>
-  );
-}
