+++ src/components/navbar/LoggedOutNavbar.jsx
@@ -1,4 +1,4 @@
-import Logo from '../assets/logo-without-bg.png';
+import Logo from '../../assets/logo-without-bg.png';
 import { useNavigate, useLocation } from 'react-router-dom';
 
 export default function LoggedOutNavbar() {
@@ -8,10 +8,6 @@ export default function LoggedOutNavbar() {
 
   const handleSignInClick = () => {
     navigate('/login?from=roadmap');
-  };
-
-  const handleSignUpClick = () => {
-    navigate('/signup');
   };
 
   const showSignIn = !pathname.includes('/login');
