+++ src/pages/LoginPage.jsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from 'react';
 
-import Navbar from '../components/Navbar';
+import Navbar from '../components/navbar/Navbar';
 import LoginCard from '../components/LoginCard';
 import analytics from '../services/posthogService';
 import { BackgroundIconCloud } from '../components/BackgroundIconCloud';
